refactor: split exerciseDb.ts into domain-focused modules
Split the 511-line exerciseDb.ts into four modules:
- exerciseTemplateDb.ts (template CRUD + search)
- workoutDb.ts (workout lifecycle + exercise management)
- exerciseSetDb.ts (set CRUD + copy logic)
- exerciseStatsDb.ts (1RM, PB, volume, history)

exerciseDb.ts becomes a barrel re-exporting everything, so
all existing consumers continue to work unchanged.
